+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js"
+    "start": "node ./server/index.js"
   },
   "engines": {
     "node": "14.x"
@@ -23,11 +23,9 @@
     "express": "^4.17.2",
     "express-flash": "^0.0.2",
     "express-session": "^1.17.2",
-    "fetch": "^1.1.0",
     "method-override": "^3.0.0",
     "mongodb": "^4.3.1",
     "mongoose": "^6.2.0",
-    "node-fetch": "^3.1.1",
     "passport": "^0.5.2",
     "passport-local": "^1.0.0"
   },
